.shiftDictionary is a Shift Dictionary
.shiftDictionary now also store shift information, not just the coworkers attending for each shift.
The 'Shift Info' and the "Coworker Info" are stored together under the same key in the .shiftDictionary.
I retrieve the "Coworker Info" from a key by splitting it at the end of the 'Shift Info' using the "Shift Info" 'location' variable, and taking the rest minus some newlines.
Because of this the resulting terminal print and .ics file generation are completely unchanged by this modification in the dictionary structure.

The "Shift Info" stored here is not currently utilized to generate the .ics file, since all the shift info is given in a single request.
Again, the "Coworker Info" is pulled from here, since it requires a request for every single shift, and since past shifts don't change, there's no sense in requesting them any more than once.
Using the dictionary to fulfill "Shift Info" to the .ics file may be a feature in the future, but at best, all it would do is reduce the (singular) Teamworx response in the amount of objects it carries. This would likely even be a significant reduction in objects, which is neat and all, but the size of the transfer isn't exactly big in the first place.
I may also have another, better idea for past shifts instead involving the "My Shifts Worked" tab of Teamworx, we'll see though.